endeavour_api/requirements.txt
asgiref
certifi
dj-database-url
Django
django-ckeditor
django-cors-headers
django-filter
django-js-asset
django-lockout
Django-Verify-Email
djangorestframework
djangorestframework-simplejwt
elastic-transport
elasticsearch
geographiclib
geopy
idna
install
pygeohash
python-dateutil
python-dotenv
pytz
requests
sib-api-v3-sdk
six
sqlparse
urllib3